Split common.h
